Wrap calls to the ns macro with stack-void
Don't want to use some arbitrary syntax to otherwise define namespaces, but also don't want to have to wrap calls to ns with (do (ns ...) :gershwin.core/stack-void). Identify when the ns macro is being called and do this wrapping automatically, before Clojure gets a chance to evaluate the forms.